doc: move TPR information to its own file, reorganize sections
TPR is deprecated and gone in K8S 1.8.  To cleanup
the README, move these details to a new file and provide
a link with note about its deprecation but compatability...

While at it, pull some of the CRD section and configuration
field definition portions earlier in the document so that they
are with the sections which introduce the topic.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>