hotplug, shim: load delegates for hotplug / hot-unplug
When adding a new interface / removing an existing one multus
cannot load the delegates for all the networks present in the CNCF
networks annotations - it must process just a single network: the
one being added or removed.

When a network / interface is added, we need to load the delegate
from the CNCF `networks` annotation; while a network / interface is
being removed, we need to load the delegate from the CNCF
`network-status` annotation (since the network is already removed
from the `networks` annotation).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>