Ensure deviceID is set when a resource is requested
If a NAD has the resourceName annotation set, then it means that the
subsequent CNI calls should provide the deviceID for that particular
resource.

Previously, we were doing best effort to get the deviceID from kubelet.
With this patch, we guarantee the deviceID will be passed in the CNI
call and fail if we can't find any deviceID.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>